requirements.txt
annotated-types    0.7.0      
anyio              4.12.1     
asgiref            3.11.0     
certifi            2025.11.12 
charset-normalizer 3.4.4      
click              8.3.1
cloudpickle        3.1.2
colorama           0.4.6
contourpy          1.3.1
cycler             0.12.1
dask               2025.12.0
distro             1.9.0
Django             5.2.9
exceptiongroup     1.3.1
fonttools          4.61.0
fsspec             2025.12.0
greenlet           3.3.0
h11                0.16.0
httpcore           1.0.9
httpx              0.28.1
idna               3.11
importlib_metadata 8.7.1
jieba              0.42.1
jiter              0.12.0
joblib             1.5.3
kiwisolver         1.4.9
locket             1.0.0
matplotlib         3.10.7
mkl_fft            2.1.1
mkl_random         1.3.0
mkl-service        2.5.2
mysqlclient        2.2.7
numpy              2.2.5
openai             2.15.0
packaging          25.0
pandas             2.3.3
partd              1.4.2
pillow             12.0.0
pip                25.3
pydantic           2.12.5
pydantic_core      2.41.5
PyExecJS           1.5.1
PyMySQL            1.1.2
pyparsing          3.2.5
python-dateutil    2.9.0.post0
pytz               2025.2
PyYAML             6.0.3
requests           2.32.5
scikit-elm         0.21a0
scikit-learn       1.7.2
scipy              1.15.3
setuptools         80.9.0
six                1.17.0
sniffio            1.3.1
snownlp            0.12.3
SQLAlchemy         2.0.45
sqlparse           0.5.5
StandardScaler     0.5
threadpoolctl      3.6.0
toolz              1.1.0
tqdm               4.67.1
typing_extensions  4.15.0
typing-inspection  0.4.2
tzdata             2025.3
urllib3            2.6.2
wheel              0.45.1
wordcloud          1.9.4
zipp               3.23.0
